Try with overloads

+++ inst/include/dust/gpu/distr/binomial.hpp
@@ -49,7 +49,7 @@ template <typename T>
 __device__
 inline float btrs(rng_state_t<T>& rng_state, float n, float p) {
   // This is spq in the paper.
-  const float stddev = sqrtf(n * p * (1 - p));
+  const float stddev = std::sqrt(n * p * (1 - p));
 
   // Other coefficients for Transformed Rejection sampling.
   const float b = 1.15f + 2.53f * stddev;
@@ -59,7 +59,7 @@ inline float btrs(rng_state_t<T>& rng_state, float n, float p) {
   const float r = p / (1 - p);
 
   const float alpha = (2.83f + 5.1f / b) * stddev;
-  const float m = floorf((n + 1) * p);
+  const float m = std::floor((n + 1) * p);
 
   float draw;
   while (true) {
@@ -67,7 +67,7 @@ inline float btrs(rng_state_t<T>& rng_state, float n, float p) {
     float v = device_unif_randf(rng_state);
     u = u - 0.5;
     float us = 0.5 - fabsf(u);
-    float k = floorf((2 * a / us + b) * u + c);
+    float k = std::floor((2 * a / us + b) * u + c);
 
     // Region for which the box is tight, and we
     // can return our calculated value This should happen
@@ -86,11 +86,11 @@ inline float btrs(rng_state_t<T>& rng_state, float n, float p) {
     // This deviates from Hormann's BRTS algorithm, as there is a log missing.
     // For all (u, v) pairs outside of the bounding box, this calculates the
     // transformed-reject ratio.
-    v = logf(v * alpha / (a / (us * us) + b));
+    v = std::log(v * alpha / (a / (us * us) + b));
     float upperbound =
-      ((m + 0.5) * logf((m + 1) / (r * (n - m + 1))) +
-       (n + 1) * logf((n - m + 1) / (n - k + 1)) +
-       (k + 0.5) * logf(r * (n - k + 1) / (k + 1)) +
+      ((m + 0.5) * std::log((m + 1) / (r * (n - m + 1))) +
+       (n + 1) * std::log((n - m + 1) / (n - k + 1)) +
+       (k + 0.5) * std::log(r * (n - k + 1) / (k + 1)) +
        stirling_approx_tail(m) + stirling_approx_tail(n - m) -
        stirling_approx_tail(k) - stirling_approx_tail(n - k));
     if (v <= upperbound) {
